requirements.txt
Django==5.2.3
djangorestframework==3.15.1
langchain==0.2.17
langchain-community==0.2.1
langchain-ollama==0.3.3
# langchain-openai==0.1.6
numpy==1.26.4
pymongo==4.13.1
python-dotenv==1.1.0
sentence-transformers==4.1.0
scikit-learn==1.7.0  # For cosine similarity
tqdm==4.67.1  # For progress bars
# dotenv==1.0.1  # For environment variable management
# pip install drf-spectacular==0.28.0
# pip install django-cors-headers